% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_station_GUI.R
\name{find_station_GUI}
\alias{find_station_GUI}
\title{Find a Historical Climate Data station (GUI)}
\usage{
find_station_GUI(stations = NULL)
}
\arguments{
\item{stations}{optional `data.frame` of stations such as a search result from `find_station()`.}
}
\value{
none
}
\description{
A function to launch a shiny web app to search for Historical Climate Data stations.
}
\examples{
\dontrun{station_explorer()}
}
\author{
Conor I. Anderson
}

