% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stations.R
\name{map_stations}
\alias{map_stations}
\title{Map canadaHCD stations on an interactive map}
\usage{
map_stations(station, zoom)
}
\arguments{
\item{station}{character; one or more station ID numbers to show on the map.}

\item{zoom}{numeric; the level to zoom the map to.}
}
\description{
Show the stations of interest on an interactive map using Leaflet. Zoom levels are guessed based on an RStudio plot window.
}
\examples{
# Make a map of all the stations named "Yellowknife".
map_stations(find_station(name = "Yellowknife"))
# Make a map of all stations within 50km of Toronto Station 5051.
map_stations(find_station(target = 5051, dist = 0:50))
}
\author{
Conor I. Anderson
}
