% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_audit.R
\name{quick_audit}
\alias{quick_audit}
\title{Perform a quick audit for missing values at a given station}
\usage{
quick_audit(datain, variables, by = "year", report = "pct",
  reverse = FALSE)
}
\arguments{
\item{datain}{A \code{tbl_df} containing the data to process}

\item{variables}{numeric or character; by default, all variables will be included. Pass a numeric vector to specify columns to include, or pass a character vector to try to match column names.}

\item{by}{character; whether values should be reported annually (\code{by = "year"}), or monthly (\code{by = "month"})}

\item{report}{character; whether values should be reported as percentage missing (\code{report = "pct"}) or as number of missing values (\code{report = "n"})}

\item{reverse}{Boolean; if \code{TRUE}, will show percentage present instead of percentage missing (only applies if \code{report = "pct"})}
}
\description{
Returns a tibble with the percentage or number of missing values at a station for a given year or year-month and variables.
}
\examples{

city <- hcd_daily(5051, 1981:2010)
quick_audit(city, "MaxTemp", by = "year", reverse = TRUE)
quick_audit(city, "MaxTemp", by = "month", report = "n")

}
\author{
Conor I. Anderson
}

