% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{hcd_daily}
\alias{reexports}
\alias{hcd_download}
\alias{reexports}
\alias{hcd_hourly}
\alias{reexports}
\alias{hcd_monthly}
\alias{reexports}
\alias{hcd_url}
\alias{reexports}
\alias{read_hcd}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{canadaHCD}{\code{\link[canadaHCD]{hcd_daily}}, \code{\link[canadaHCD]{hcd_download}}, \code{\link[canadaHCD]{hcd_hourly}}, \code{\link[canadaHCD]{hcd_monthly}}, \code{\link[canadaHCD]{hcd_url}}, \code{\link[canadaHCD]{read_hcd}}}
}}

