% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_data.R
\name{get_station_data}
\alias{get_station_data}
\title{Get latest Environment and Climate Change Canada Station Data}
\usage{
get_station_data(assume_yes = FALSE)
}
\description{
Get latest Environment and Climate Change Canada Station Data
}
\keyword{internal}
