% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_audit.R
\name{quick_audit}
\alias{quick_audit}
\title{Perform a quick audit for missing values at a given station}
\usage{
quick_audit(datain, variables, reverse = FALSE)
}
\arguments{
\item{datain}{A \code{\link{tbl_df}} containing the data to process}

\item{variables}{numeric or character; By default, all variables will be included. Pass a numeric vector to specify columns to include, or pass a character vector to try to match column names.}

\item{reverse}{Boolean; if true, will show proportion present instead of proportion missing.}
}
\description{
Returns a tibble with the proportion of missing values at a station for a given year and variables.
}
\examples{

\dontrun{
city <- hcd_daily(5051, 1981:2010)
quick_audit(city, "MaxTemp", reverse = TRUE)
}

}
\author{
Conor I. Anderson
}

